backend/requirements.txt
fastapi
uvicorn
pydantic
groq==0.4.2
python-dotenv
httpx==0.24.1
